requirements.txt
jieba==0.39
nltk==3.3
numpy==1.14.5
gensim==3.4.0
Keras==2.2.0
progressbar2==3.38.0
scikit_learn==0.19.1
